==> scripts/r/00_libs.R <==
# Load libraries --------------------------------------------------------------

# General packages
library("here")
library("fs")
library("purrr")
library("dplyr")
library("tidyr")
library("forcats")
library("stringr")
library("readr")
library("readxl")
library("writexl")
library("glue")

# Plotting packages
library("ggplot2")
library("ganttrify")
library("contributoR")
library("ggbeeswarm")
library("ggrepel")
library("patchwork")
library("ggridges")
library("extrafont")

# Modeling packages
library("brms")
library("tidybayes")
library("bayestestR")

# -----------------------------------------------------------------------------
